+++ custom_components/virtual_mrt_top/select.py
@@ -32,11 +32,10 @@ from .const import (
     CONF_RADIANT_TYPE,
     RADIANT_TYPES,
     CONF_IS_RADIANT,
-CONF_DEVICE_TYPE,
+    CONF_DEVICE_TYPE,
     TYPE_AGGREGATOR,
+    get_device_info,
 )
-from .device_info import get_device_info
-
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -51,11 +50,12 @@ async def async_setup_entry(
     device_info = await get_device_info({(DOMAIN, entry.entry_id)}, config[CONF_NAME])
     store_key = f"{STORAGE_KEY}_{entry.entry_id}"
     store = Store(hass, STORAGE_VERSION, store_key)
-    entities: List[VirtualProfileSelect | VirtualRadiantTypeSelect] = [VirtualProfileSelect(hass, entry, device_info, store)]
+    entities: List[VirtualProfileSelect | VirtualRadiantTypeSelect] = [
+        VirtualProfileSelect(hass, entry, device_info, store)
+    ]
     if config.get(CONF_IS_RADIANT, False):
         entities.append(VirtualRadiantTypeSelect(hass, entry, device_info))
     async_add_entities(entities)
-
 
 
 # --- NEW CLASS ADDED: VirtualRadiantTypeSelect ---